+++ CHANGELOG.md
@@ -4,6 +4,7 @@
    - artboard dimensions control inside editing window
    - fill and stroke updates on element selection
    - polygon and polyline visual editing
+   - line visual editing
  - 0.8.1
    - fill and stroke controls inside editing window
    - hints